order-independent name resolution \o/ :D the power of CTE
one of the coolest pieces of code I've ever written I think. definitions
can be placed however is desired, and are executed in whatever order is
necessary for them to make sense.

this also automatically implemented compile-time evaluation for `def`
constants! basically the fluent runtime is shaping up to be insanely
intuitive in a way I did not predict.