ssa functions and nicely rendering ssa
basically I realized that the previous idea of the `Block` simply wasn't
going to be enough to represent programs well. now, the functionality of
the previous `Block` abstraction is split into `Func`s, which contain a
namespace for const, local, and block references, and `Block`s, which
contain linear operations. this maintains the linear structure of the
original idea while making control flow easy to implement and also (not
coincidentally) tracking much closer with the planned QBE and LLVM
backends.

in parallel, I also added kritzler rendering for the SSA IR, which is
very pretty :) (and convenient)