execute slice construction from the address of an array!
still need to resurrect slices, but the internal execution works :)

this also gives me an idea of what compiling other kinds of structured
data will require, and it should be very straightforward. it's all about
keeping the conventions crystal clear so that all of my generated code
can rely on the other pieces of itself.