`Result(T, E)` and `Project` (NOT FUNCTIONAL)
basically this is a rewrite of the context system, which required
sweeping changes to everything that touches the frontend. the result is
a lot of simplified code, and some really nice patterns that are much
harder to break. very nice!

currently lexing is still somewhat broken, but the contextual messaging
and kritzler rendering is working extremely nicely :)