parallel parsing
this works surprisingly well. I could continue to hack on this parser
and get something reasonable at the end, but there are a few problems
that I think would be solved more nicely with a more "do a lot of work
and then error with tons of context and backtrack" kind of approach.

gotta remember that parsing here is not about performance whatsoever, I
in fact barely care at all about performance as long as it doesn't get
in the realm of noticeably slow, and as long as that speed cost results
in error messaging that is nice to read, and a parser that is nice to
develop for