fast table-based parsing :)
I thought this optimization would add complexity, but it actually seems
as though it has simplified my parsing code, which is absolutely sick.
parsing is now measured in milliseconds and submilliseconds which goes
above and beyond my expectations. very happy :)

all that is left is implementing flags, which hopefully won't be a
problem? crossing my fingers