multiple rule per prefix/infix in parser; crystallize slice type/collection syntax
instead of using `{}` be purely associative, `{}` now also stands in for
arrays as a generalized collection syntax. given the semantics of dict
mappings, I think this makes a lot of sense.

this now frees up `[]` to be used for slice types `[]T` and many
pointers `[*]T`, which are bits of syntax I really like in zig