fluent Reprs
data representation in ssa has been encoded through high-level types,
but this was lacking and doesn't quite match the model of SSA. where
high-level fluent is more symbolic, focusing on symbolic notation,
constness, and (planned) move semantics, low-level fluent is a
procedural, copy-based language. Repr reflects this.

immediate goals with Repr are to:
1. narrow the functionality of SSA operations
2. make SSA code verification easy
3. handle low-level details of structured data and the stack layout
  - this will also make implementing closures seamless, since SSA sees
    a function's stack and fluent structured data through the same lens
4. make qbe/llvm targets seamless
5. narrow the responsibilities of Types