compiling basic TExprs with new IdMap system
this is super leaky and crashes everywhere at the moment, but it has
already forced me to think more clearly about the internal data
representation of TExprs and how it relates to Type/Value/Repr which is
incredible, honestly.

I also realized I've inadvertently created the ability to introspect on
all of the live objects in a fluent program, which is crazy.