Object.Interface + represent Expr in fluent :)
this is an absurdly important commit. basically I got the first piece of
the puzzle working, TExpr has been resurrected (ha) in the form of Expr,
which is an in-fluent representation of the typed AST.

this commit also importantly includes `struct` and `variant` types,
which mean that programming in fluent is about to get much more high
level.