self referential types
basically just consisted of tracking stacked selves for identifying a
zig type, along with transitioning over to using com.IdMap for the
backing TypeId to Type mappings!

this commit is still very broken due to Reprs requiring the same
treatment, but this is an incredibly minor commit for something I
thought might be an incredibly major problem.