add Image to Env; write too many TODOs
everything is coming together, but there is going to be a bit to manage
here with integration.

the central task is going to be removing Value from everything and replacing
it with Ptr. this only really has a single challenging surface area,
which is Object.

anyways, this is a pretty exciting commit. fluent is falling into place
and I am excited about it.